[bug fix][string] fix the wrong usage of snprintf, and use function 'memcpy' to replace 'strcpy'
The snprintf() will write at most n-1 of the characters printed into the output string
(the n'th character then gets the terminat-ing `\0');

if the return value is greater than or equal to the n argument,
the string was too short and some of the printed characters were discarded.

The output is always null-terminated.