replace PF_INET to AF_INET
Reference to socket manual page:
http://man7.org/linux/man-pages/man2/socket.2.html

NOTES:
The manifest constants used under 4.x BSD for protocol families are
PF_UNIX, PF_INET, and so on, while AF_UNIX, AF_INET, and so on are
used for address families.  However, already the BSD man page
promises: "The protocol family generally is the same as the address
family", and subsequent standards use AF_* everywhere.